+++ classes/Controllers/MediaController.php
@@ -6,6 +6,7 @@ namespace Grav\Plugin\FlexObjects\Controllers;
 
 use Grav\Common\Form\FormFlash;
 use Grav\Common\Grav;
+use Grav\Common\Page\Media;
 use Grav\Common\Page\Medium\Medium;
 use Grav\Common\Session;
 use Grav\Common\Uri;
@@ -204,9 +205,14 @@ class MediaController extends AbstractController
 
         $name = $this->getPost('name');
         $settings = $object->getBlueprint()->schema()->getProperty($name);
-
-        $media = $object->getMedia();
-        $folder = $settings['folder'] ?? Utils::url($media->path()) ?: null;
+        $fieldFolder = $settings['folder'] ?? null;
+        if ($fieldFolder) {
+            // Custom media.
+            $media = new Media($fieldFolder, []);
+        } else {
+            // Object media.
+            $media = $object->getMedia();
+        }
 
         $available_files = [];
         $metadata = [];
@@ -230,13 +236,15 @@ class MediaController extends AbstractController
 
         // Peak in the flashObject for optimistic filepicker updates
         $pending_files = [];
-        $sessionField  = base64_encode($this->getGrav()['uri']->url());
-        $flash         = $this->getSession()->getFlashObject('files-upload');
+        $sessionField = base64_encode($this->getGrav()['uri']->url());
+        $flash = $this->getSession()->getFlashObject('files-upload');
+        $folder = $media->getPath() ?: null;
 
         if ($flash && isset($flash[$sessionField])) {
             foreach ($flash[$sessionField] as $field => $data) {
                 foreach ($data as $file) {
-                    if (\dirname($file['path']) === $folder) {
+                    $test = \dirname($file['path']);
+                    if ($test === $folder) {
                         $pending_files[] = $file['name'];
                     }
                 }